fix the column of size

+++ database/migrations/2024_04_23_021651_create_food_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->text('picture');
             $table->enum('food_type', ['berat', 'kemasan']);
             $table->enum('food_category', ['makanan', 'minuman']);
-            $table->decimal('size');
+            $table->integer('size');
             $table->string('barcode_number')->nullable();
             $table->timestamps();
         });
